// src/module/keypad/enums/keypad_enums.ts
export enum KeypadSymbol {
    ASH = "ash",
    AT = "at",
    B = "b",
    BALLOON = "balloon",
    BLACK_STAR = "black star",
    BUTT = "butt",
    C_WITH_DOT = "c_with_dot",
    KITTY = "kitty",
    COPYRIGHT = "copyright",
    CURLY_H = "curly h",
    DOUBLE_K = "double k",
    DRAGON = "dragon",
    EURO = "euro",
    FORK = "fork",
    HALF_THREE = "half_three",
    LAMBDA = "lamda",
    LETTER_N = "letter n",
    LIGHTNING = "lightning",
    OMEGA = "omega",
    PARAGRAPH = "paragraph",
    PIG_TAIL = "pigtail",
    QUESTION = "question",
    REVERSE_C = "reverse c",
    SIX = "six",
    SMILEY = "smiley",
    STITCH = "stitch",
    WHITE_STAR = "white star"
}